Move upgrade to people and remove print statements
